+++ tests/Integration/RestAPI/Firmware/RestAPI.spec.ts
@@ -3,6 +3,7 @@ import RestApi from "../../../../src/RestApi/RestApi";
 import { accountOpts, opts, RUN_INTEGRATION_TESTS } from "../helper";
 import IFirmwareVersion, { IFirmwareVersionUpdatable } from '../../../../src/RestApi/Firmware/Version/IFirmwareVersion';
 import FirmwareVersion from '../../../../src/RestApi/Firmware/Version/FirmwareVersion';
+import { createReadableStream } from '../../../Unit/RestApi/Applet/Version/File/helper';
 
 const allowedTimeout = 10000;
 const api = new RestApi(opts, accountOpts);
@@ -14,8 +15,14 @@ describe('RestAPI - FirmwareVersion', () => {
 		'applicationType': 'webos',
 		'version': '04.01.74',
 		'createdAt': new Date('2017-05-24T08:56:52.550Z'),
-		'hash': '8e9c3ded774d7b021be452570e0aba10',
-		'confirmed': false, // TODO: ask misak
+		'uploaded': false,
+		'files': [
+			{
+				'content': createReadableStream(''),
+				'hash': '8e9c3ded774d7b021be452570e0aba10',
+				'size': 12345,
+			},
+		],
 	};
 
 	before (function() {
@@ -47,20 +54,26 @@ describe('RestAPI - FirmwareVersion', () => {
 		await api.firmwareVersion.create({
 			applicationType: `webos`,
 			version: `04.01.74`,
-			hash: `8e9c3ded774d7b021be452570e0aba10`,
+			files: [
+				{
+					content: createReadableStream('i am very beautiful stream of bytes'),
+					hash: `8e9c3ded774d7b021be452570e0aba10`,
+					size: 12345,
+				},
+			],
 		});
 		should(true).true();
 
 	}).timeout(allowedTimeout);
 
-	it('should set firmwareVersion confirmed', async function () {
+	it('should set firmwareVersion uploaded', async function () {
 
-		const update: IFirmwareVersionUpdatable = { confirmed: true };
+		const update: IFirmwareVersionUpdatable = { uploaded: true };
 		await api.firmwareVersion.set('webos', '04.01.74', update);
 		const firmwareVersions = await api.firmwareVersion.list();
 		firmwareVersions.forEach((fwv: IFirmwareVersion) => {
 			if (fwv.uid === firmwareVersion.uid) {
-				should.equal(fwv.confirmed, update.confirmed);
+				should.equal(fwv.uploaded, update.uploaded);
 			}
 		});